server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
boto3==1.34.0
pdf2image==1.16.3
Pillow==10.1.0 